chore: added layoyuts config prettier ts and vite
