+++ .prettierrc.json
@@ -9,7 +9,7 @@
   "jsxSingleQuote": false,
   "plugins": ["@trivago/prettier-plugin-sort-imports"],
   "importOrderGroupNamespaceSpecifiers": true,
-  "importOrderParserPlugins": ["typescript", "jsx", "tsx"],
+  "importOrderParserPlugins": ["typescript", "jsx"],
   "importOrderSortSpecifiers": true,
   "importOrderSeparation": true,
   "importOrder": [
