+++ src/entities/stack/index.ts
@@ -1,3 +1,3 @@
 export * from "./view";
-export * as boardLib from "./constants";
+export * from "./constants";
 export * from "./model";
