+++ src/entities/user/index.ts
@@ -1,3 +1,2 @@
-export * from "./ui";
-export * as userLib from "./lib";
-export * as userModel from "./model";
+export * from "./view";
+export * from "./model";
