+++ src/entities/workspace/index.ts
@@ -1,1 +1,1 @@
-export * as workspaceModel from "./model";
+export * from "./model";
