+++ src/features/add-entity/index.ts
@@ -1,3 +1,1 @@
-export * as addEntityLib from "./lib";
-export * as addEntityModel from "./model";
-export * from "./ui";
+export * from "./view";
