--- src/features/add-entity/lib/helpers.ts
@@ -1,1 +1,0 @@
-export {};
