--- src/features/add-entity/lib/index.ts
@@ -1,2 +1,0 @@
-export * as models from "./models";
-export * as helpers from "./helpers";
