+++ src/features/add-entity/view.tsx
@@ -1,11 +1,19 @@
 import { memo } from "react";
+import type { ChangeEventHandler, FormEventHandler } from "react";
 
 import { Button, CloseXButton } from "~/shared/ui/button";
 import { InputArea } from "~/shared/ui/input";
 
-import { IAddEntityProps } from "./lib/models";
+interface AddEntityProps {
+  editable: boolean;
+  onChange: ChangeEventHandler<HTMLTextAreaElement>;
+  onReset: FormEventHandler<HTMLFormElement>;
+  onSubmit: FormEventHandler<HTMLFormElement>;
+  value: string;
+  buttonCaption: string;
+}
 
-export const AddEntity = memo<IAddEntityProps>(
+export const AddEntity = memo<AddEntityProps>(
   ({ editable, onChange, onReset, onSubmit, value, buttonCaption }) => {
     return (
       <form onReset={onReset} onSubmit={onSubmit} className="flex flex-col gap-4 px-4 py-1">
