+++ src/features/add-list/index.ts
@@ -1,3 +1,1 @@
-export * from "./ui";
-export * as addListLib from "./lib";
-export * as addListModel from "./model";
+export * from "./view";
