+++ src/features/add-list/model.ts
@@ -1,11 +1,1 @@
-import { createEvent } from "effector";
-
-export const actions = {};
-
-const xxx = createEvent();
-
-xxx.watch(console.log);
-
-xxx.watch(() => {
-  console.log("xxx");
-});
+export {};
