--- src/features/auth/social-auth-button/lib/index.ts
@@ -1,2 +1,0 @@
-export * as models from "./models";
-export * as helpers from "./helpers";
