+++ src/features/auth/social-auth-button/model.ts
@@ -1,0 +1,18 @@
+import type { ButtonHTMLAttributes } from "react";
+
+export type SocialButtonAuthServices =
+  | "google"
+  | "facebook"
+  | "apple"
+  | "figma"
+  | "twitter"
+  | "dribbble";
+
+export interface SocialButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  social: SocialButtonAuthServices;
+  noCaption?: boolean;
+  pending?: boolean;
+  theme?: SocialButtonTheme;
+}
+
+export type SocialButtonTheme = "brand" | "color" | "colorWithBrand";
