+++ src/features/board/add-to-favorite/index.ts
@@ -1,1 +1,1 @@
-export * from "./ui";
+export * from "./view";
