+++ src/features/board/add-to-favorite/view.tsx
